+++ spec/rubocop/config_store_spec.rb
@@ -20,6 +20,8 @@ describe RuboCop::ConfigStore do
       .to receive(:load_file) { |arg| RuboCop::Config.new(arg) }
     allow(RuboCop::ConfigLoader)
       .to receive(:merge_with_default) { |config| "merged #{config.to_h}" }
+    allow(RuboCop::ConfigLoader)
+      .to receive(:default_configuration) { 'default config' }
   end
 
   describe '.for' do
@@ -47,6 +49,15 @@ describe RuboCop::ConfigStore do
         expect(RuboCop::ConfigLoader).to receive(:configuration_file_for).once
         expect(RuboCop::ConfigLoader).to receive(:configuration_from_file).once
         config_store.for('file1')
+      end
+
+      context 'when --force-default-config option is specified' do
+        it 'uses default config without searching for config path' do
+          expect(RuboCop::ConfigLoader).to_not receive(:configuration_file_for)
+          expect(RuboCop::ConfigLoader).to_not receive(:configuration_from_file)
+          config_store.force_default_config!
+          expect(config_store.for('file1')).to eq('default config')
+        end
       end
     end
   end